+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Handle errors on bad BibTeX data in New Source and return a more useful message
 - Handle errors on negative width/height at api/v1/images/123/ocr
 - Prevent bad OCR requests from SQED image task
+- Valid/Invalid Taxon Name facet value wasn't being used in the Exact case
 
 ### Changed